LayoutTests/imported/w3c/web-platform-tests/css/css-color/color-mix-basic-001.tentative-expected.txt


FAIL From blue to red at 0.5 assert_equals: Colors should match for color-mix(in srgb, blue, red  50%) expected "rgb(128, 0, 128)" but got "color(srgb 0.5 0 0.5)"
FAIL From blue to green at 0.5 assert_equals: Colors should match for color-mix(in srgb, blue, green  50%) expected "rgb(0, 64, 128)" but got "color(srgb 0 0.2509804 0.5)"
FAIL From rgb(255, 0, 0, .2) to red at 0.5 assert_equals: Colors should match for color-mix(in srgb, rgb(255, 0, 0, .2), red  50%) expected "rgba(255, 0, 0, 0.6)" but got "color(srgb 1 0 0 / 0.6)"
FAIL From blue to red at 0.9 assert_equals: Colors should match for color-mix(in srgb, blue, red  90%) expected "rgb(230, 0, 26)" but got "color(srgb 0.9 0 0.1)"
FAIL From blue to red at 0 assert_equals: Colors should match for color-mix(in srgb, blue, red  0%) expected "rgb(0, 0, 255)" but got "color(srgb 0 0 1)"
FAIL From currentColor to white at 0.5 assert_not_equals: Should parse color-mix(in srgb, currentColor, white  50%) got disallowed value ""
FAIL From currentColor to rgba(0, 0, 0, .5) at 0.5 assert_not_equals: Should parse color-mix(in srgb, currentColor, rgba(0, 0, 0, .5)  50%) got disallowed value ""
